+++ src/adt/set.inc
@@ -109,55 +109,6 @@ set_create(const struct fsm_alloc *a,
 	return set;
 }
 
-static struct set *
-set_create_singleton(const struct fsm_alloc *a,
-	int (*cmp)(const void *a, const void *b), item_t *item)
-{
-	struct set *s;
-
-	assert(cmp != NULL);
-
-	s = f_malloc(a, sizeof *s);
-	if (s == NULL) {
-		return NULL;
-	}
-
-	s->a = f_malloc(a, sizeof *s->a);
-	if (s->a == NULL) {
-		f_free(a, s);
-		return NULL;
-	}
-
-	s->alloc = a;
-	s->a[0] = item;
-	s->i = s->n = 1;
-	s->cmp = cmp;
-
-	return s;
-}
-
-static struct set *
-set_copy(const struct set *set)
-{
-	struct set *s;
-	s = malloc(sizeof *s);
-	if (s == NULL) {
-		return NULL;
-	}
-
-	s->cmp = set->cmp;
-	s->a = malloc(set->i * sizeof s->a[0]);
-	if (s->a == NULL) {
-		free(s);
-		return NULL;
-	}
-
-	s->i = s->n = set->i;
-	memcpy(s->a, set->a, s->n * sizeof s->a[0]);
-
-	return s;
-}
-
 static item_t *
 set_add(struct set *set, item_t *item)
 {
@@ -214,31 +165,6 @@ set_add(struct set *set, item_t *item)
 }
 
 static void
-set_remove(struct set *set, const item_t *item)
-{
-	size_t i;
-
-	assert(set != NULL);
-	assert(set->cmp != NULL);
-	assert(item != NULL);
-
-	if (set_empty(set)) {
-		return;
-	}
-
-	i = set_search(set, item);
-	if (set->cmp(&item, &set->a[i]) == 0) {
-		if (i < set->i) {
-			memmove(&set->a[i], &set->a[i + 1], (set->i - i - 1) * (sizeof *set->a));
-		}
-
-		set->i--;
-	}
-
-	assert(!set_contains(set, item));
-}
-
-static void
 set_free(struct set *set)
 {
 	assert(set != NULL);
@@ -246,147 +172,5 @@ set_free(struct set *set)
 
 	free(set->a);
 	free(set);
-}
-
-static size_t
-set_count(const struct set *set)
-{
-	assert(set != NULL);
-	assert(set->a != NULL);
-
-	return set->i;
-}
-
-static void
-set_clear(struct set *set)
-{
-	assert(set != NULL);
-	assert(set->a != NULL);
-
-	set->i = 0;
-}
-
-static int
-set_cmp(const struct set *a, const struct set *b)
-{
-	assert(a != NULL);
-	assert(a->a != NULL);
-	assert(b != NULL);
-	assert(b->a != NULL);
-
-	if (a->i != b->i) {
-		return a->i - b->i;
-	}
-
-	return memcmp(a->a, b->a, a->i * sizeof *a->a);
-}
-
-static int
-set_equal(const struct set *a, const struct set *b)
-{
-	assert(a != NULL);
-	assert(a->a != NULL);
-	assert(b != NULL);
-	assert(b->a != NULL);
-
-	if (a->i != b->i) {
-		return 0;
-	}
-
-	return 0 == memcmp(a->a, b->a, a->i * sizeof *a->a);
-}
-
-static item_t *
-set_first(const struct set *set, struct set_iter *it)
-{
-	assert(set != NULL);
-	assert(set->a != NULL);
-	assert(it != NULL);
-
-	if (set_empty(set)) {
-		it->set = NULL;
-		return NULL;
-	}
-
-	it->i = 0;
-	it->set = set;
-
-	return it->set->a[it->i];
-}
-
-static item_t *
-set_firstafter(const struct set *set, struct set_iter *it, const item_t *item)
-{
-	size_t i;
-	int r;
-
-	assert(set != NULL);
-	assert(set->cmp != NULL);
-	assert(set->a != NULL);
-	assert(it != NULL);
-
-	if (set_empty(set)) {
-		it->set = NULL;
-		return NULL;
-	}
-
-	i = set_search(set, item);
-	r = set->cmp(&item, &set->a[i]);
-	assert(i <= set->i - 1);
-
-	if (r >= 0 && i == set->i - 1) {
-		it->set = NULL;
-		return NULL;
-	}
-
-	it->i = i;
-	if (r >= 0) {
-		it->i++;
-	}
-
-	it->set = set;
-	return it->set->a[it->i];
-}
-
-static item_t *
-set_next(struct set_iter *it)
-{
-	assert(it != NULL);
-
-	it->i++;
-	if (it->i >= it->set->i) {
-		return NULL;
-	}
-
-	return it->set->a[it->i];
-}
-
-static item_t *
-set_only(const struct set *set)
-{
-	assert(set != NULL);
-	assert(set->n >= 1);
-	assert(set->i == 1);
-	assert(set->a[0] != NULL);
-
-	return set->a[0];
-}
-
-static int
-set_hasnext(const struct set_iter *it)
-{
-	assert(it != NULL);
-
-	return it->set && it->i + 1 < it->set->i;
-}
-
-static const item_t *
-set_array(const struct set *set)
-{
-	if (set == NULL) {
-		return NULL;
-	}
-
-	return (const item_t *) set->a;
 }
