+++ tests/set/set2.c
@@ -34,9 +34,6 @@ int main(void) {
 	assert(set_add(s, &a));
 	assert(set_add(s, &a));
 
-	set_remove(s, &a);
-	assert(!set_contains(s, &a));
-
 	set_free(s);
 
 	return 0;
